pyproject.toml
[project]
name = "RedirectService"
version = "0.1.0"
description = ""
authors = [
    {name = "Taras",email = "taras.stepeniuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "shortuuid (>=1.0.13,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
responses = "^0.25.6"
factory-boy = "^3.3.3"

